+++ src/Identity/IdentityServer/RequestValidators/TwoFactorAuthenticationValidator.cs
@@ -16,7 +16,7 @@ using Bit.Core.Utilities;
 using Duende.IdentityServer.Validation;
 using Microsoft.AspNetCore.Identity;
 
-namespace Bit.Identity.IdentityServer;
+namespace Bit.Identity.IdentityServer.RequestValidators;
 
 public interface ITwoFactorAuthenticationValidator
 {
